Update gallery.js
Wrapped addImgToCarousel in a checker to solve if it exists or not. Added buildCarousel function to the second half of the conditional. Moved the display for gallery detail container to the top of the conditional. changed buildCarousel container id in order to match the previous id. Changed a method from set attribute to style